[RSDK-4488]Rename packages in sensor controlled base (#2913)
