+++ components/base/register/register.go
@@ -5,6 +5,6 @@ import (
 	// register bases.
 	_ "go.viam.com/rdk/components/base/agilex"
 	_ "go.viam.com/rdk/components/base/fake"
-	_ "go.viam.com/rdk/components/base/sensorbase"
+	_ "go.viam.com/rdk/components/base/sensorcontrolled"
 	_ "go.viam.com/rdk/components/base/wheeled"
 )
