+++ components/base/sensorcontrolled/sensorcontrolled_test.go
